Refactor javascript
