+++ src/utils/outputListFromTags.js
