+++ src/utils/splitOnHyphen.js
